% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{auto_t_test}
\alias{auto_t_test}
\title{Run t test
Run t test with two vector after homogeneity of variance test.}
\usage{
auto_t_test(v1, v2, ...)
}
\arguments{
\item{v1}{a numeric vector.}

\item{v2}{a numeric vector.}

\item{...}{Other arguments of \code{\link[stats]{t.test}}}
}
\description{
Run t test
Run t test with two vector after homogeneity of variance test.
}
