% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{auto_wilcox_test}
\alias{auto_wilcox_test}
\title{Run wilcox test
Run wilcox test with two vector.}
\usage{
auto_wilcox_test(v1, v2, ...)
}
\arguments{
\item{v1}{a numeric vector.}

\item{v2}{a numeric vector.}

\item{...}{Other arguments of \code{\link[stats]{wilcox.test}}}
}
\description{
Run wilcox test
Run wilcox test with two vector.
}
