% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{batch_combat}
\alias{batch_combat}
\title{Remove batch effect of RNA-seq data}
\usage{
batch_combat(
  expressionMatrix,
  batch,
  covariate = NULL,
  par.prior = TRUE,
  prior.plots = FALSE,
  ...
)
}
\arguments{
\item{expressionMatrix}{Genomic measure matrix (dimensions probe x sample) - for example, expression matrix.
It is better to use a log transformed data as input, and there must be no negative value in the data,
because it will generate NaN value.}

\item{batch}{Batch covariate (only one batch allowed)}

\item{covariate}{A data frame contain the covariates instead in. Each column must be factor with the same sample
order with the data. For example, If the batch of different experimental platforms is removed and the
differences of normal and cancer are calculated, then the normal and a CANCER samples in each experiment
are variables that are interested, this data are generally variables that are interested in subsequent
research. When batch correction, you need to consider the difference in quantity distribution in each batch (unbalanced)}

\item{par.prior}{(Optional) TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used}

\item{prior.plots}{(Optional) TRUE give prior plots with black as a kernel estimate of the empirical batch effect density and red as the parametric}

\item{...}{Other arguments of \code{\link[sva]{ComBat}}}
}
\value{
data A probe x sample genomic measure matrix, adjusted for batch effects.
}
\description{
Remove batch effect of RNA-seq data by combat
}
\note{
If the result of \code{Combat} contain negative values, replace by zero
}
\examples{
\dontrun{

library(bladderbatch)
data(bladderdata)
pheno = pData(bladderEset)
expressionMatrix = exprs(bladderEset)
batch = pheno$batch
covariate = data.frame(cancer = as.factor(pheno$cancer))
correctMatrix <- FunctionCombat(expressionMatrix, batch, covariate = covariate)

}
}
\seealso{
\url{https://www.biostars.org/p/196430/}
\url{https://support.bioconductor.org/p/73266/}
\url{https://www.biostars.org/p/173684/}
}
