% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_with_meta.R
\name{bind_with_meta}
\alias{bind_with_meta}
\title{Combine data with metadata}
\usage{
bind_with_meta(
  ...,
  .names_meta = NULL,
  .prefix = NULL,
  .fun_pre = NULL,
  .fun_post = NULL,
  .delim_name = "_"
)
}
\arguments{
\item{...}{Data frames to combine. The names of the data frames should
contain the metadata.}

\item{.names_meta}{Character vector of names of the metadata columns. If
\code{NULL}, the metadata parsing is skipped.}

\item{.prefix}{Prefix to be removed from the target names to retrieve the
metadata.}

\item{.fun_pre, .fun_post}{Functions to apply to the data before and after
combining.}

\item{.delim_name}{Delimiter used in the names of the data frames to separate
the metadata.}
}
\value{
A data frame with the combined data and metadata.
}
\description{
This is a wrapper around \code{dplyr::bind_rows()} that allows to combine
multiple data frames with metadata. The metadata is extracted from the
names of the data frames. The names of the data frames should be of the
form \code{prefix_meta1_meta2_...} given the delimiter is set as \verb{_}. The
metadata columns are then extracted and converted to the appropriate
types.
}
