% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_full.R
\name{call_full}
\alias{call_full}
\title{Create a call to a function with its arguments}
\usage{
call_full(.fn)
}
\arguments{
\item{.fn}{The function to call.}
}
\value{
A call to the function with its arguments.
}
\description{
This is basically a wrapper around \code{\link[rlang:call2]{rlang::call2()}} that allows you to
extract the arguments from a function and pass them to \code{\link[rlang:call2]{rlang::call2()}}
without having to type them out.
}
