% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cautiously.R
\name{cautiously}
\alias{cautiously}
\title{Cautiously evaluate an expression}
\usage{
cautiously(.f, otherwise = NULL)
}
\arguments{
\item{.f}{A function to modify. See \code{\link[=as_function]{as_function()}} for details.}

\item{otherwise}{A value to return if the expression throws an error.}
}
\value{
A function that evaluates the expression and returns the result
or the default value if an error is thrown.
}
\description{
This function is useful for when you want to evaluate an expression, but
you want to catch any errors and return a default value instead. Note the
error message will be printed as a warning as the name of the function
suggests.
}
