% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{check_log2}
\alias{check_log2}
\title{Check whether log2 transformed}
\usage{
check_log2(mat)
}
\arguments{
\item{mat}{a numeric matrix or a numeric vector for log2 transformation}
}
\description{
\code{check_log2} can check data whether have a log2 transformation
}
\examples{
check_log2(1:10)
check_log2(matrix(log2(1:10)))

}
