% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell.R
\name{dirichlet_regression}
\alias{dirichlet_regression}
\title{Dirichlet multinomial regression to detect changes in cell frequencies}
\usage{
dirichlet_regression(counts, covariates, formula)
}
\arguments{
\item{counts}{a matrix with cell number, a samples x cell types matrix}

\item{covariates}{additional data to use in the regression}

\item{formula}{the formula of regression}
}
\value{
a matrix of cell proportion
}
\description{
This function use Dirichlet multinomial regression calculate the cell proportion of single cell data
}
\examples{
\dontrun{
  counts = do.call(cbind, tapply(seur@data.info$orig.ident, seur@ident, table))
  covariates = data.frame(condition=gsub('[12].*', '', rownames(counts)))
  res = dirichlet_regression(counts, covariates, counts ~ condition)
}

}
