% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helper.R
\name{download_kegg_geneset}
\alias{download_kegg_geneset}
\title{Downlaod KEGG pathway}
\usage{
download_kegg_geneset(species = "hsa", OrgDb = "org.Hs.eg.db")
}
\arguments{
\item{species}{supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{OrgDb}{OrgDb to convert gene ID}
}
\value{
A \code{tibble} object contain pathway informations
}
\description{
Download KEGG pathway and gene list
}
