% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{fill_by_constant}
\alias{fill_by_constant}
\title{Impute the NA value in matrix}
\usage{
fill_by_constant(
  mat,
  method = c("median", "mean", "min", "max"),
  MARGIN = c("row", "column"),
  contant_value = NULL
)
}
\arguments{
\item{mat}{a matrix object}

\item{method}{The method used to fill NA value. Can be
\describe{
\item{\code{median}:}{Use the row or column median value to fill NA}
\item{\code{mean}:}{Use the row or column mean value to fill NA}
\item{\code{min}:}{Use the global min value to fill NA}
\item{\code{max}:}{Use the global max value to fill NA}
}}

\item{MARGIN}{choose the direction two get the mean or median value to fill NA. It is only useful when \code{method} is \code{mean} or \code{median}}

\item{contant_value}{the contant value to fill NA. Default is NULL. If you set this argument, all NA value in the matrix will fill by it}
}
\value{
return a matrix after fill the NA
}
\description{
Impute the NA value in a matrix by different method
}
\examples{
\dontrun{
 a <- matrix(c(1:5, NA), ncol = 2)
 fill_by_constant(mat = a)
 fill_by_constant(mat = a, contant_value = 1)

}
}
