% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{fill_by_knn}
\alias{fill_by_knn}
\title{Fill NA by KNN method}
\usage{
fill_by_knn(mat, ...)
}
\arguments{
\item{mat}{A matrix}

\item{...}{Other arguments of \code{\link[impute]{impute.knn}}}
}
\value{
A matrix with imputed values
}
\description{
A function to fill na value using nearest neighbor averaging.
}
\examples{
\dontrun{
  ww <- matrix(data = 1:20, ncol = 4, nrow = 5)
  ww[13] <- NA
  fill_by_knn(ww)
}

}
