% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{fill_by_normal_distribution}
\alias{fill_by_normal_distribution}
\title{Impute NA by normal distribution}
\usage{
fill_by_normal_distribution(mat, width = 0.3, downshift = 1.8, seed = 1)
}
\arguments{
\item{mat}{A matrix}

\item{width}{Scale factor for the standard deviation of imputed distribution relative to the sample standard deviation.}

\item{downshift}{Down-shifted the mean of imputed distribution from the sample mean, in units of sample standard deviation.}

\item{seed}{Random seed. Defalut: 1}
}
\value{
return a matrix after fill the NA
}
\description{
Impute NA values by random numbers drawn from a normal distribution that has a down-shifted mean
and shrunken standard deviation from the sample distribution. This is meant to be similar to imputation
in the Perseus software.
}
\examples{
\dontrun{
  ww <- matrix(data = 1:20, ncol = 4, nrow = 5)
  ww[13] <- NA
  fill_by_normal_distribution(ww)
}

}
