% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funny.R
\name{funny_christmas_tree}
\alias{funny_christmas_tree}
\title{Marry Christmas}
\usage{
funny_christmas_tree(bulbs = TRUE, snow = TRUE, star = TRUE, bg.col = 1)
}
\arguments{
\item{bulbs}{Whether plot bulbs.}

\item{snow}{Whether plot snow.}

\item{star}{Whether plot star.}

\item{bg.col}{Same as \code{bg} in \code{\link[graphics]{par}}}
}
\description{
Create a Christmas tree
}
