% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Collected general color palette}
\usage{
get_color_palette(
  n,
  alpha = 1,
  theme = c("maftools_domain_col", "maftools_mutation1", "maftools_mutation2",
    "maftools_snv", "cellchat", "palette1")
)
}
\arguments{
\item{n}{number of colors to generate}

\item{alpha}{factor modifying the opacity alpha; typically between 0 to 1}

\item{theme}{color from different color theme}
}
\value{
return a color vector
}
\description{
Many beautiful color palette collected by Erjie.
}
\examples{
get_color_palette(n = 5, alpha = 0.7, theme = 'cellchat')

}
