% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{na_ratio}
\alias{na_ratio}
\title{NA ratio in data}
\usage{
na_ratio(x)
}
\arguments{
\item{x}{a matrix with any type data}
}
\value{
a value of NA ratio
}
\description{
This function get the NA ratio in the whole data
}
