% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{na_ratio_col}
\alias{na_ratio_col}
\title{NA ratio of each column in data}
\usage{
na_ratio_col(x)
}
\arguments{
\item{x}{a matrix with any type data}
}
\value{
a vector value of NA ratio in each column
}
\description{
This function get the NA ratio in the column of data
}
