% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{na_ratio_row}
\alias{na_ratio_row}
\title{NA ratio of each row in data}
\usage{
na_ratio_row(x)
}
\arguments{
\item{x}{a matrix with any type data}
}
\value{
a vector value of NA ratio in each row
}
\description{
This function get the NA ratio in the row of data
}
