% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_venn}
\alias{plot_venn}
\title{Make Venn Diagram from a list}
\usage{
plot_venn(x, ...)
}
\arguments{
\item{x}{A list of vectors (e.g., integers, chars), with each component corresponding to a separate circle in the Venn diagram}

\item{...}{other argument of \link[VennDiagram:venn.diagram]{venn.diagram}}
}
\description{
Make Venn Diagram base on \link[VennDiagram:venn.diagram]{venn.diagram}, you can save it by \code{ggsave}
}
\examples{
\dontrun{
  kp <- plot_venn(list(a = sample(letters[1:10], 20, replace = T),
                       b = sample(letters[1:10], 25, replace = T)),
                  fill=c("#CC79A7", "#56B4E9"),
                  col=c("#D55E00", "#0072B2"),
                  cat.col=c("#D55E00", "#0072B2"))
}
}
\seealso{
\link[VennDiagram:venn.diagram]{venn.diagram}
}
\author{
Erjie Zhao
}
