% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{prism_theme}
\alias{prism_theme}
\title{Prism themes}
\usage{
prism_theme(
  base_family = "sans",
  base_size = 14,
  base_fontface = "bold",
  base_line_size = base_size/14,
  base_rect_size = base_size/14,
  axis_text_angle = 0,
  border = FALSE
)
}
\arguments{
\item{base_family}{\code{string}. Base font family, default is \code{"sans"}.}

\item{base_size}{\code{numeric}. Base font size, given in \code{"pt"}.}

\item{base_fontface}{\code{string}. Base font face, default is \code{"bold"}.}

\item{base_line_size}{\code{numeric}. Base linewidth for line elements}

\item{base_rect_size}{\code{numeric}. Base linewidth for rect elements}

\item{axis_text_angle}{\code{integer}. Angle of axis text in degrees.
One of: \verb{0, 45, 90, 270}.}

\item{border}{\code{logical}. Should a border be drawn around the plot?
Clipping will occur unless e.g. \code{coord_cartesian(clip = "off")} is used.}
}
\value{
Returns a list-like object of class \emph{theme}.
}
\description{
A collection of ggplot2 themes that use palettes which mirror the
colour schemes available in GraphPad Prism.
}
