% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{Read all excel sheets}
\usage{
read_all_sheets(file, sheets = NULL, merge = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{file}{Path to the xls/xlsx file.}

\item{sheets}{name or number to read, if NULL read all sheets.}

\item{merge}{Whether merge the data, default FALSE.}

\item{verbose}{Whether print useful message.}

\item{...}{Other arguments of \code{\link[readxl]{read_excel}}}
}
\description{
Read specific sheet or all sheets in an excel.
}
