% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyeval.R
\name{select_list}
\alias{select_list}
\title{Tidy select for list}
\usage{
select_list(.l, ...)
}
\arguments{
\item{.l}{A \code{\link[=list]{list()}} object.}

\item{...}{One or more unquoted expressions separated by commas.}
}
\value{
A list with the selected elements.
}
\description{
A tidy select interface for lists. See \code{\link[tidyselect:eval_select]{tidyselect::eval_select()}}
for details.
}
