% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_combine_with_meta}
\alias{tar_combine_with_meta}
\title{Static aggregation with metadata}
\usage{
tar_combine_with_meta(
  name,
  cols_meta,
  ...,
  prefix = NULL,
  fun_pre = identity,
  fun_post = identity
)
}
\arguments{
\item{name}{Name of the target. Could be a string or symbol.}

\item{cols_meta}{Character vector of column names to store the retrieved
metadata.}

\item{...}{One or more target objects or list of target objects. See
\code{\link[tarchetypes:tar_combine]{tarchetypes::tar_combine()}}.}

\item{prefix}{Prefix to be removed from the target names to retrieve the
metadata. Defaults is the same as \code{name}.}

\item{fun_pre}{Function to apply to each target before combining.}

\item{fun_post}{Function to apply to the combined data frame.}
}
\value{
A target object. See \code{\link[tarchetypes:tar_combine]{tarchetypes::tar_combine()}}.
}
\description{
Like \code{\link[tarchetypes:tar_combine]{tarchetypes::tar_combine()}} except it adds metadata columns to the
combined data frame.
}
\details{
This function is useful for combining data frames. For \code{\link[=list]{list()}} targets,
you should pass a function to convert the list to a data frame in the
\code{fun_pre} argument. Metadata is retrieved from the names of the targets.
}
