% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transpose_dgRMatrix}
\alias{transpose_dgRMatrix}
\title{Transpose dgRMatrix to dgCMatrix}
\usage{
transpose_dgRMatrix(dgRMatrix_obj)
}
\arguments{
\item{dgRMatrix_obj}{a \code{dgRMatrix} object}
}
\value{
A dgCMatrix that is the transposed dgRMatrix
}
\description{
Transpose dgRMatrix to dgCMatrix
}
