% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows_meta.R
\name{bind_rows_meta}
\alias{bind_rows_meta}
\title{Combine multiple data frames by row with metadata from names}
\usage{
bind_rows_meta(
  ...,
  .names_meta = NULL,
  .patterns_meta = NULL,
  .delim = "_",
  .prefix = NULL,
  .fun_pre = NULL,
  .fun_post = NULL
)
}
\arguments{
\item{...}{Data frames to combine. Typically, they are named appropriately to
extract the metadata.}

\item{.names_meta, .patterns_meta}{Character vectors of names and patterns to
extract the metadata from the names of the data frames. The \code{.names_meta}
is used to name the metadata columns and \code{.patterns_meta} is used to
extract the values. If \code{.names_meta} is \code{NULL}, the metadata parsing is
skipped. If \code{.patterns_meta} is \code{NULL}, they are assumed to be \code{".*"}.}

\item{.delim}{Delimiter used in the names of the data frames to separate
different pieces of metadata.}

\item{.prefix}{Prefix to be removed from the target names to retrieve the
metadata.}

\item{.fun_pre, .fun_post}{Functions to apply to the data before and after
combining. Note they should return values of \code{\link[=data.frame]{data.frame()}} class.}
}
\value{
A data frame with the combined data and metadata.
}
\description{
This is a wrapper around \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} that allows to combine multiple
data frames with metadata. The metadata is extracted from the names of the
data frames. The names of the data frames should be of the form
\code{prefix_meta1_meta2_...} given the delimiter is set as \verb{_}. The metadata
columns are then extracted and converted to the appropriate types.
}
